[Common] - Fix - Finalizing import module entities migration +++

+++ RR-Common/src/main/java/com/scor/rr/domain/DefaultAdjustmentRegionPerilEntity.java
@@ -9,7 +9,7 @@ import java.util.Objects;
 @IdClass(DefaultAdjustmentRegionPerilEntityPK.class)
 public class DefaultAdjustmentRegionPerilEntity {
     private int fkDefaultAdjustmentId;
-    private int fkRegionPerilId;
+    private long fkRegionPerilId;
     private String includedExcluded;
     private DefaultAdjustmentEntity defaultAdjustment;
     private RegionPerilEntity regionPeril;
@@ -26,7 +26,7 @@ public class DefaultAdjustmentRegionPerilEntity {
 
     @Id
     @Column(name = "RegionPerilId", nullable = false)
-    public int getFkRegionPerilId() {
+    public long getFkRegionPerilId() {
         return fkRegionPerilId;
     }
